Break out into smaller files to support testing
